third_party/WebKit/LayoutTests/html/marquee/marquee-direction-attribute-should-be-case-insensitive.html

<!DOCTYPE html>
<marquee direction="DOwN">The text should move down.</marquee>
<marquee direction="uP">The text should move up.</marquee>
<marquee direction="RiGHt">The text should move right.</marquee>
<marquee direction="leFT">The text should move left.</marquee>
<script src="../../resources/js-test.js"></script>
<script>
var marquees = document.querySelectorAll('marquee');
for (i = 0; i < marquees.length; ++i) {
  var marquee = marquees[i];
  var computedStyle = getComputedStyle(marquee);
  switch (marquee.getAttribute('direction').toLowerCase()) {
  case 'down':
  case 'up':
    shouldBeEqualToString("computedStyle.whiteSpace", "normal");
    break;
  case 'right':
  case 'left':
  default:
    shouldBeEqualToString("computedStyle.whiteSpace", "nowrap");
  }
}
</script>
