third_party/blink/web_tests/html/marquee/marquee-direction-attribute-should-be-case-insensitive-expected.txt

PASS computedStyle.whiteSpace is "normal"
PASS computedStyle.whiteSpace is "normal"
PASS computedStyle.whiteSpace is "nowrap"
PASS computedStyle.whiteSpace is "nowrap"
PASS successfullyParsed is true

TEST COMPLETE

The text should move down. The text should move up. The text should move right. The text should move left.
